Removed embark/disembark. Merged settle/visit/station commands
